requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
defusedxml==0.7.1
diff-match-patch==20200713
Django
django-allauth==0.49.0
django-crispy-forms==1.14.0
django-import-export==2.7.1
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
MarkupPy==1.14
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.0.1
pycparser==2.21
PyJWT==2.3.0
python-decouple==3.6
python3-openid==3.2.0
PyYAML==6.0
razorpay==1.3.0
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
tablib==3.2.0
urllib3==1.26.8
whitenoise==6.0.0
xlrd==2.0.1
xlwt==1.3.0
